+++ apps/ddaas/src/app/[locale]/(content)/legal/layout.tsx
@@ -35,8 +35,7 @@ export default async function Layout({
       enabled: false,
     },
     themeSwitch: {
-      enabled: true,
-      mode: 'light-dark-system',
+      mode: 'light-only',
     },
   };
 
@@ -62,6 +61,7 @@ export default async function Layout({
           tree: legalSource.getPageTree(locale),
           sidebar: { enabled: false },
           searchToggle: { enabled: false },
+          themeProvider: false,
         }}
       >
         {children}